fix(FormRenderer): Honour isSubmitting prop in Wizard
The `Wizard` component doesn't use the `FormTemplate`, and so wasn't displaying a button with a
loading spinner when `isSubmitting` was true.

We use context to track whether or not the form is submitting, which means we don't re-render the
entire form every time `isSubmitting` is changed, which means the wizard doesn't "reset" to the
first step which was a pretty horrible user experience!